Support compound key combinations in GlobalKey
Compound keys can now be formed by embedding a KeyCommandSet inside
of a CompoundKeyCommand.  The set is only enabled for a short time
(250 ms) after the activation key on the CompoundKeyCommand is hit
by the user.  If the timeout expires, the set is disabled.  If a
key is selected, the set is disabled.

Signed-off-by: Shawn O. Pearce <sop@google.com>