BROKEN - Test ordering of RegexFindReplace rules
This test is actually broken right now.  The 2nd rule replaces
the content in the anchor of the 1st rule, making the result
actually have an href inside of an href, e.g.:

  <a href="">gwtexpui <a href="">issue 9918</a></a>

Which is perhaps technically correct given our use of replaceAll
but not what the user meant to have happen.

Change-Id: Ia17b771a4bd8486e77a5b2b2ea9c80d803fb54dd
Signed-off-by: Shawn O. Pearce <sop@google.com>
Reviewed-by: Brad Larson <bklarson@gmail.com>