Update patch set 1
Ugh, this is a good catch.  I'm not sure about the best solution... ideally, we should probably scan in order and avoid scanning substrings which we've already matched, but I doubt that is very straightforward with javascript.

We could do something slick with a recursive function - find a match for the first item in the list, then recursively call with the substring from the start to the first match, and when that returns if it found a match make another recursive call from the end of that match to the start of the current match... I'm not sure what your opinion is on recursion...  This would be building a lot of objects if there are lots of matches, so it might degrade performance by an unacceptable margin.

Patch-set: 1