Update patch set 1
I think what we'd want to do is always match the longest match possible at any position.  Which is what any regex engine will do by default if we toss all of them into a single alternate list.

So if we create a new super pattern by taking all of the RegexFindReplace.match values and putting them together with "|" operators we'd get what we need to always pick the longest possible match.

Then we'd want to perform the replace by executing code at each match site that redoes each individual pattern one at a time against the complete matched substring, but using ^$ at the start/end of each of those patterns to ensure they match the entire thing.  When we find the actual matching pattern, we use its replace on that segment, and continue the super pattern on the main string.

The only sane way to do this is run the matcher looking for all matches, literally copying the sections which don't match, but working on the sections that do.

In Java we can do this with java.util.regex.Matcher, and its actually how String.replaceAll is implemented.  Unfortunately GWT doesn't have a replacement for java.util.regex, so you can't use java.util.regex directly from GWT code.  But JavaScript does have the equivalent functionality and GWT's String.replaceAll is depending on it:

https://developer.mozilla.org/en/Core_JavaScript_1.5_Reference/Global_Objects/RegExp

If you want to see GWT's code look at the source:

http://code.google.com/p/google-web-toolkit/source/browse/releases/1.7/user/super/com/google/gwt/emul/java/lang/String.java

I propose we write our own tiny regex wrapper that has two modes, one using java.util.regex.Matcher in Java, and one using JSNI to access RegExp in JavaScript.    Right now we'd have to test the JSNI based one by hand, but it should be pretty thin and almost a direct port of the java.util.regex.Matcher wrapper.

Patch-set: 1