Fix diappearance of text in CopyableLabel when clicking on it
When the user clicks on a CopyableLabel the displayed label is
replaced by an input box. In some cases the displayed input box has a
visible length of 0 so that the contained text is not displayed. This
happens when the text for the CopyableLabel is not set via the
constructor but with the setText(...) method. The problem is that the
member variable that holds the visible length is not updated when the
setText(...) method is invoked.

Bug: issue 1176
Signed-off-by: Edwin Kempin <edwin.kempin@sap.com>