Update patch set 2
Patch Set 2:

Not sure if this is related to this change or if this was already an issue before, but URL's that contain a Change-Id are incorrectly formatted.

https://gerrit-review.googlesource.com/#/q/I0dbdc5ff7804450d8b5c8a424a8cb1b0059b4029,n,z

This is because we are making a link out of the URL and in addition linkify the Change-Id.

Patch-set: 2
Reviewer: Gerrit User 1003935 <1003935@173816e5-2b9a-37c3-8a2e-48639d4f1153>
Label: Code-Review=0